+++ src/main/java/lotto/controller/LottoController.java
@@ -9,9 +9,9 @@ import lotto.model.dto.WinningResultDto;
 import lotto.model.service.LottoComparator;
 import lotto.model.service.lottopublisher.LottoPublisher;
 import lotto.model.service.prizecalculator.PrizeCalculator;
-import lotto.model.domain.PublishedLotto;
+import lotto.model.domain.lotto.PublishedLotto;
 import lotto.model.domain.PurchasePrice;
-import lotto.model.domain.WinningLotto;
+import lotto.model.domain.lotto.WinningLotto;
 import lotto.model.dto.PublishedLottoDto;
 import lotto.util.InputUtil;
 import lotto.util.LottoNumberUtil;
