+++ src/main/java/lotto/model/service/LottoComparator.java
@@ -2,9 +2,9 @@ package lotto.model.service;
 
 import java.util.List;
 import lotto.model.domain.CompareResult;
-import lotto.model.domain.Lotto;
-import lotto.model.domain.PublishedLotto;
-import lotto.model.domain.WinningLotto;
+import lotto.model.domain.lotto.Lotto;
+import lotto.model.domain.lotto.PublishedLotto;
+import lotto.model.domain.lotto.WinningLotto;
 
 public class LottoComparator {
